refactor - changed bodeyWeight in Infant.java to Double data type from Long and as well modified the setters, getters, request and response where it's used. Added changeset to modify the database as well.

+++ src/main/java/org/lamisplus/modules/pmtct/domain/dto/InfantDto.java
@@ -21,7 +21,7 @@ public class InfantDto implements Serializable
     private String ancNo;
     private String infantOutcomeAt18Months;
     private String personUuid;
-    private Long bodyWeight;
+    private Double bodyWeight;
     private String ctxStatus;
     private InfantArvDto infantArvDto;
     private InfantPCRTestDto infantPCRTestDto;
